Add files via upload
This is a morse code project which I made just for fun. Honestly I didn't know about morse code and how to communicate with it. I have watched a lot of spy movies where it is used and I had that fire in me that I want to learn this. After I learned the morse code, I decided to make a simple program which converts morse code to english and english to morse code.